Dev in dev

+++ ogsr_engine/Layers/xrRender/HOM.cpp
@@ -104,7 +104,7 @@ void CHOM::Load			()
 		rT.adjacent[0]	= (0xffffffff==adjacency[3*it+0])?((occTri*) (-1)):(m_pTris+adjacency[3*it+0]);
 		rT.adjacent[1]	= (0xffffffff==adjacency[3*it+1])?((occTri*) (-1)):(m_pTris+adjacency[3*it+1]);
 		rT.adjacent[2]	= (0xffffffff==adjacency[3*it+2])?((occTri*) (-1)):(m_pTris+adjacency[3*it+2]);
-		rT.flags		= clT.dummy;
+		rT.flags		= u32(clT.dummy);
 		rT.area			= Area	(v0,v1,v2);
 		if (rT.area<EPS_L)	{
 			Msg	("! Invalid HOM triangle (%f,%f,%f)-(%f,%f,%f)-(%f,%f,%f)",VPUSH(v0),VPUSH(v1),VPUSH(v2));
