+++ ogsr_engine/Layers/xrRender/HW.cpp
@@ -550,7 +550,7 @@ void	CHW::updateWindowProps	(HWND m_hWnd)
 			// desktop.
 
 			RECT			m_rcWindowBounds;
-			float fYOffset = 0.f;
+			int fYOffset = 0;
 			static const bool bCenter = !strstr(Core.Params, "-no_center_screen");
 
 			if(bCenter){
@@ -573,7 +573,7 @@ void	CHW::updateWindowProps	(HWND m_hWnd)
 									DevPP.BackBufferHeight );
 			};
 
-			AdjustWindowRect		(	&m_rcWindowBounds, dwWindowStyle, FALSE );
+			AdjustWindowRect		(	&m_rcWindowBounds, DWORD(dwWindowStyle), FALSE );
 
 			SetWindowPos			(	m_hWnd, 
 										HWND_NOTOPMOST,
