+++ ogsr_engine/Layers/xrRenderPC_R3/r3_rendertarget_phase_smap_D.cpp
@@ -34,8 +34,8 @@ void	CRenderTarget::phase_smap_direct		(light* L, u32 sub_phase)
 	else
 	{
 		D3D_VIEWPORT VP					=	{L->X.D.minX,L->X.D.minY,
-			(L->X.D.maxX - L->X.D.minX) , 
-			(L->X.D.maxY - L->X.D.minY) , 
+			UINT(L->X.D.maxX - L->X.D.minX) , 
+			UINT(L->X.D.maxY - L->X.D.minY) ,
 			0,1 };
 		//CHK_DX								(HW.pDevice->SetViewport(&VP));
 		HW.pDevice->RSSetViewports(1, &VP);
